+++ composeApp/src/commonMain/kotlin/ui/trips/core/mapper/TripFilterParcelizeMapper.kt
@@ -3,16 +3,16 @@ package ui.trips.core.mapper
 import core.domain.cities.model.City
 import profile.domain.transport.TransportBrand
 import profile.domain.transport.TransportType
-import trips.filter.presentation.TripFilterScreenState
 import trips.all.domain.models.TripsFilter
+import trips.filter.presentation.TripFilterScreenState
 import ui.core.presentation.models.CityParcelize
 import ui.trips.core.models.TripsFilter as UiTripsFilter
 
 fun TripFilterScreenState.toTripsFilterParcelize() = UiTripsFilter(
     autoType = selectedState.selectedAutoType,
     autoModel = selectedState.selectedAutoModel,
-    fromPriceTrip = selectedState.selectedFromPriceTrip,
-    toPriceTrip = selectedState.selectedToPriceTrip,
+    fromPriceTrip = selectedState.selectedFromPriceTrip.toInt(),
+    toPriceTrip = selectedState.selectedToPriceTrip.toInt(),
     fromCity = selectedState.selectedFromCity?.let { CityParcelize(it.id, it.name) },
     toCity = selectedState.selectedToCity?.let { CityParcelize(it.id, it.name) },
     tripDate = selectedState.selectedTripDate,
